Update to swf-tree 0.7
This commit updates the SWF parser and type definitions to version 0.7. The `swf-tree` and `swf-parser` versions should match from now on.

This update includes minor fixes to font tags. It does not seem impact the code of `flashback` but it's good to keep the dependencies up-to-date. Most of the work has gone into the testing infrastructure: the libraries now share about a hundred test samples defined in the [open-flash-db](https://github.com/open-flash-db/) repositories.

Also, there are changelogs now 🎉
- https://github.com/open-flash/swf-tree/blob/master/CHANGELOG.md#070-2019-05-21
- https://github.com/open-flash/swf-parser/blob/master/CHANGELOG.md#070-2019-05-21