Update to swf-types 0.11
The crate `swf-tree` got renamed to `swf-types` as it's a more general name.

The highlight of this update is the improved parser reliability: invalid input now returns an error instead of panicking. This is tested with new tests and fuzzing. If the parser ever panics, you should report it as a bug. The parser also has a more friendly Rust API.
Besides the improved parser, there were some tweaks to the types to better support invalid tags, undocumented `DoAbc` tags and bitmaps. On the Rust side, `serde` is now an optional feature of the SWF types, it should help consumers such as `flashback` that do not care about strict serialization.

- https://github.com/open-flash/swf-types/blob/master/CHANGELOG.md#0110-2020-02-05
- https://github.com/open-flash/swf-parser/blob/master/CHANGELOG.md#0110-2020-02-05